pyproject.toml
[project]
name = "inbox-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromadb>=1.4.0",
    "colorlog>=6.10.1",
    "fastapi>=0.128.0",
    "flask>=3.1.2",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.3",
    "gradio>=6.2.0",
    "ipykernel>=7.1.0",
    "langchain-huggingface>=1.2.0",
    "litellm>=1.80.11",
    "lxml>=6.0.2",
    "numpy>=2.4.0",
    "openai>=2.14.0",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.2.0",
    "streamlit>=1.52.2",
    "tenacity>=9.1.2",
    "tqdm>=4.67.1",
]
